Added speed powerup and updated Powerup
Update Powerup abstract class.

Do not kill() when the player picks up the powerup. This causes the
powerup to get off of the powerups list, which means that we can't
update it after the player picks it up. We need to be able to update
since some powerups have longer effects.

To get around this, in the render() method. Check if the powerup has
been picked up, if not, then render image. If it has, then continue.