Abstracted GamePlay states. Updated TimeStop to stop spawns also
For timestop to stop spawns, it now takes in a gamestate as an argument
to the constructor. TimeStop then sets the spawnOn field to false, until
after a delay, at which time it resets spawnOn to true.

GamePlayState is an abstract class that is the template for both single
and double player game play.